+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - bugfix for pane contents restoration
 - expand tilde char `~` if used with `@resurrect-dir`
 - do not save empty trailing lines when pane content is saved
+- do not save pane contents if pane is empty (only for 'save pane contents'
+  feature)
 
 ### v2.4.0, 2015-02-23
 - add "tmux-test"